Split functions into appropriate files
